Move wallet related stuff to AccountProvider
