+++ app/containers/AccountProvider/actions.js
@@ -29,6 +29,8 @@ export const PROXY_EVENTS = 'acebusters/AccountProvider/PROXY_EVENTS';
 
 export const CONTRACT_EVENTS = 'acebusters/AccountProvider/CONTRACT_EVENTS';
 
+export const WALLET_LOADED = 'acebusters/App/WALLET_LOADED';
+
 // Note: refer to  https://developer.mozilla.org/en-US/docs/Web/API/WebSocket#Ready_state_constants
 export const READY_STATE = {
   CONNECTING: 0,
@@ -196,4 +198,8 @@ export function contractEvent(event, proxy) {
 
 export function contractEvents(events, proxy) {
   return { type: CONTRACT_EVENTS, payload: events, meta: { proxy } };
+}
+
+export function walletLoaded(wallet) {
+  return { type: WALLET_LOADED, payload: wallet };
 }
